feat: use postCreateHook for create client special logic
