+++ package.json
@@ -29,6 +29,7 @@
   },
   "homepage": "https://soundworks.dev",
   "dependencies": {
+    "@ircam/sc-utils": "^1.13.0",
     "chalk": "^5.3.0",
     "commander": "^11.1.0",
     "compare-versions": "^6.1.1",
